requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv

django-debug-toolbar
pytest
pytest-django
black
isort

django-crispy-forms
crispy-bootstrap5
